streamlit/requirements.txt
accelerate==0.26.1
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.3
asgiref==3.7.2
asitop==0.0.24
astrapy==0.6.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.7.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
cassandra-driver==3.28.0
cassio==0.1.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
cleo==2.1.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmdstanpy==1.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
dashing==0.1.0
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
effdet==0.4.1
einops==0.7.0
elastic-transport==8.11.0
elasticsearch==8.11.1
email_validator==2.1.1
emoji==2.11.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.4
filetype==1.2.0
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
geomet==0.2.1.post1
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-generativeai==0.3.1
googleapis-common-protos==1.61.0
greenlet==3.0.3
groq==0.5.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h2==4.1.0
holidays==0.39
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.8.0
importlib-resources==6.1.1
install==1.3.5
installer==0.7.0
iopath==0.1.10
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.1
kiwipiepy==0.17.1
kiwipiepy-model==0.17.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-experimental==0.0.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langserve==0.0.51
langsmith==0.1.47
layoutparser==0.3.4
libcst==1.4.0
llvmlite==0.41.1
lxml==5.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mlx==0.9.1
mmh3==4.0.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.3
ngrok==1.0.0
nltk==3.8.1
npm==0.1.1
numba==0.58.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
open-clip-torch==2.24.0
openai==1.17.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
optional-django==0.1.0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
patsy==0.5.5
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==8.15.0
pillow==10.3.0
pillow_heif==0.16.0
pinecone-client==4.0.0
pkginfo==1.10.0
platformdirs==4.0.0
playwright==1.44.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.41
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.6
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.8.0
pydantic-settings==2.3.2
pydantic_core==2.18.1
pydeck==0.8.0
pyee==11.1.0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypandoc==1.12
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-google-places==1.4.2
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.8.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.1
selenium==4.16.0
semantic-version==2.10.0
sentence-transformers==2.6.1
sentencepiece==0.1.99
setuptools-rust==1.8.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==2.1.0
stack-data==0.6.3
stanio==0.3.0
starlette==0.37.2
statsmodels==0.14.1
streamlit==1.33.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.39.3
Tree==0.2.4
trio==0.23.2
trio-websocket==0.11.1
trove-classifiers==2024.3.25
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.13.2
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
validators==0.22.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.12
webdriver-manager==4.0.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
xattr==1.1.0
xformer==1.0.1
xlrd==2.0.1
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
